% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swildboot.R
\name{swildboott}
\alias{swildboott}
\title{Score based Wild bootstrap improvements for inference with small cluser size (MLE)}
\usage{
swildboott(model, cluster, beta = "All", type = "mammen", confid = 0.05,
  R = 400, seed = NULL)
}
\arguments{
\item{model}{A nonlinear model (glm,polr) estimated using \code{lm}. If the mdoel is estimated by OLS, please use \code{wboot} instead.}

\item{cluster}{A formula to specify the variable of cluster.}

\item{beta}{A key variable of interest. To denote the key varibale to focus on, one could save a amount of time because the algorithm will only report the outcome of key variable. The defalt value is "All" which will report all the results of every coefficents.}

\item{type}{The type of distributions for wild boostrap. Default value is Mammen. One could also use Rademacher though the Mammen distribution seems to do much better than Rademacher for small cluster size.}

\item{confid}{The size of test}

\item{R}{The number of boostrap replicates. The default number is 400.}

\item{seed}{A single value passed to \code{set.seed}.}
}
\description{
The function is used to get accurate and precise inference from small cluster sizes. Through  (restricted) wild clustered bootstrap-t to get p-values for MLE models, the function considerably reduces the probability of over rejection of the null hypotheses. The distribution used to reconstruct the residuals is Mammen or Rademacher.
}
\examples{
\dontrun{

##### glm model #####
data(wv6_equ)  # input data
# create binomial dependent variable
wv6_bi <- as.data.frame(wv6_equ)
wv6_bi$income_equ0[wv6_bi$income_equ < 6] <- 0
wv6_bi$income_equ0[wv6_bi$income_equ > 5] <- 1
glm_equ <- glm(income_equ0 ~ income + age + gender, data = wv6_equ) # glm model
swildboott(glm_equ, ~country, R = 500) # find original p values are over estimated

##### polr model #####
data(wv6_equ)  # input data
require(MASS)
wv6_equ <- as.data.frame(wv6_equ)
polr_equ <- polr(factor(income_equ) ~ income + age + gender, data = wv6_equ) # polr model
swildboott(polr_equ, ~country, R = 500) # find original p value is over estimated
}


}
\references{
Kline, Patrick and Andres Santos. 2012 "A Score Based Approach to Wild Bootstrap Inference" \emph{Journal of Econometric Methods} 2012; 1(1): 23–41
}
\author{
Jiawei Fu \email{jiawei.fu@duke.edu}
}
