% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildboott.R
\name{wildboott}
\alias{wildboott}
\title{Wild bootstrap improvements for inference with small cluser size (OLS)}
\usage{
wildboott(model, cluster, beta = "All", type = "Rademacher",
  confid = 0.05, R = 200, seed = NULL)
}
\arguments{
\item{model}{A linear model estimated using \code{lm}. If the mdoel is  nonlinear, please use \code{swboot} instead.}

\item{cluster}{A formula to specify the variable of cluster.}

\item{beta}{A key variable of interest. To denote the key varibale to focus on, one could save a amount of time because the algorithm will only report the outcome of key variable. The defalt value is 'All' which will report all the results of every coefficents.}

\item{type}{The type of distributions for wild boostrap. Default value is Rademacher. If the number of clusters is less than 12, please use 'six' instead (type='six')}

\item{confid}{The size of test}

\item{R}{The number of boostrap replicates. The default number is 200.}

\item{seed}{A single value passed to \code{set.seed}.}
}
\description{
The function is used to get accurate and precise inference from small cluster sizes. Through  (restricted) wild clustered bootstrap-t to get p-values for OLS models, the function considerably reduces the probability of over rejection of the null hypotheses.The default distribution used to reconstruct the residuals is Rademacher. When cluster size is smaller than 12, the algorithm will suggest to use "six - point" distribution instead.
}
\details{
linear model
impose H0 where H0: beta = 0
}
\examples{
\dontrun{
data(wv6_equ)  # input data
wv6_equ <- as.data.frame(wv6_equ)
model_equ <- lm(income_equ ~ income + age + gender, data = wv6_equ) # linear model
wildboott(wv6_equ, ~country, R = 250) # find original p value is over estimated
}

}
\references{
Cameron, A. Colin, Jonah B. Gelbach, and Douglas L. Miller. 2008. "Bootstrap-Based Improvements for Inference with Clustered Errors" \emph{The Review of Economics and Statistics} 90(3): 414-427.

Cameron, A. Colin and Douglas L. Miller. 2015. "A Practitioner’s Guide to Cluster-Robust Inference" \emph{The Journal of Human Resources} Vol. 50 No. 2 317-372

Webb, Matthew D. 2013. "Reworking wild bootstrap based inference for clustered errors" \emph{Queen's Economics Department Working Paper}
}
\author{
Jiawei Fu \email{jiawei.fu@duke.edu}
}
