% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wv6_equ}
\alias{wv6_equ}
\title{World Value Survey wave 6 dataset}
\format{A data frame with 558 observations and 5 variables:
\describe{
  \item{country}{code for each country}
  \item{income_equ}{the preference for income redistribution from 1 to 10}
  \item{income}{individuals' income level}
  \item{age}{individuals' age}
  \item{gender}{individuals' gender}
}}
\source{
\url{http://www.worldvaluessurvey.org/wvs.jsp}
}
\usage{
wv6_equ
}
\description{
A dataset containing variables related to preference for income equality, including country cluster, income scale, age, and individual gender.
}
\keyword{datasets}
